Extract topics main page to dedicated model
